Removed unneeded file
Added SettingsHeaderViewModel to display the Settings icon in the titlebar

Added ShellView border

Added SettingsView

Settings flyout is positioned on the right and is being toggled by the icon in the title bar

removed unused dialog service

Small ShellVM cleanup

Adds SessionManager to take care of the LogOut process. It's called by the SettingsViewModel and it stops the OmniService starts the ActivationService

Revert "removed unused dialog service"

This reverts commit 28a8b7a3dbdc859ab0b2da5f8331d5b83d579af2.

Login button is now logging you out of the application and brings up the authorization dialog.

Renamed method on ConfigurationService

Adds tests for ConfigurationService