Added Clippings View. This is a combination of 14 commits:
ClipboardHandler is now acting as a sequence of Clippings. It merges OmniClipboardHandler and LocalClipboardHandler

Added ClippingListViewModel to display the Clippings on the screen. It subscribes to ClipboardHandler and when a new Clipping arrives, it adds it to a list.

Added Source on Clippings. This can be Web or Local. OmniClipboardHandler sets it to Web and LocalClipboardHandler sets it to Local.

LocalClipboardHandler is now subscribing to WindowsClipboardHandler only once.

Added clippingViewModel

Separated the Clippings into: All, Local, Web and added some styling.

Wrapped clipping text

Renamed Web (Clippings) to Cloud (Clippings)

Added possibility to open a link in the browser

Added test to show that LocalClipboardHandler must only subscribe once to the WindowsClipboardWrapper.DataReceived's event

Refactored ClippingListViewModel in one Master viewModel and three different and simple VMs that derive from ClippingListViewModelBase

Refactoring: The three Clipping ViewModels are now using the same View using the UseViewAttribute.

Renamed file

Added tests for ClippingListViewModelBase