Wo 107 - Adds (basic) android install guide
This is a combination of 8 commits.
Transformed Incoming Clipping to Incoming SMS and put a Reply button on the notification view. This is a fakery for the filming crew

Fixes unresponsive application bug. This is a combination of 4 commits.
Changed the name of the button

Adds exit button in the SettingsViewModel. This is a combination of 3 commits.

Adds exit button in the SettingsViewModel

Removed manual binding for ApplicationService

Renamed MulticontentButtonBehavior to MulticontentButton since it's just a dependency property and not a behavior

Revert "Transformed Incoming Clipping to Incoming SMS and put a Reply button on the notification view. This is a fakery for the filming crew"

This reverts commit 835639be97180bfec336d11fc44e5500c0eba43d.

Conflicts:
	Code/Omnipaste/Notification/ClippingNotification/ClippingNotificationView.xaml

Fixed bug where the application was unresponsive

Tray icon changes depending on the status of the application.  This is a combination of 4 commits.

ContextMenuViewModel is now subscribed to OmniServce status changed observable and changes the icon accordingly

Conflicts:
	Code/Omnipaste/Omnipaste.csproj

Revert "OmniService doesn't implement IObservable<StatusChanged> directly but exposes an observable"

This reverts commit 327437af3a32f8ced04ce08cd2f6f84ed7d539a8.

Added new icons for connected and disconnected

Revert "Revert "OmniService doesn't implement IObservable<StatusChanged> directly but exposes an observable""

This reverts commit b6d339f21c6466616125caa5f1f3a61f3dc56a6c.

Application is shown the first time it is used.

Added a step on the ActivationService that checks the number the user has registered

Added AndroidInstallGuide step on the ActivationService

LoadingViewModel handles ShowAndroidGuideMessage to display the AndroidInstallationGuideView

The android guide is dismissed when you press the button