Fixes unresponsive application bug. This is a combination of 4 commits. Changed the name of the button
Adds exit button in the SettingsViewModel. This is a combination of 3 commits.

Adds exit button in the SettingsViewModel

Removed manual binding for ApplicationService

Renamed MulticontentButtonBehavior to MulticontentButton since it's just a dependency property and not a behavior

Revert "Transformed Incoming Clipping to Incoming SMS and put a Reply button on the notification view. This is a fakery for the filming crew"

This reverts commit 835639be97180bfec336d11fc44e5500c0eba43d.

Conflicts:
	Code/Omnipaste/Notification/ClippingNotification/ClippingNotificationView.xaml

Fixed bug where the application was unresponsive