WO-114 - Added IncomingSmsNotification. This is a combination of 11 commits. Added SMSMessage model
fixed tests

Added IncomingSmsNotificationViewModel with the possibility to reply to Smss

Removed unused models

fixed test

Extracted common interface for IncomingCallNotificationVM annd IncomingSmsNotificationVM and refactored NotificationViewModelFactory to use it.

Used the Content property to show the SMS message in the notification

Refactored tests

Refactored the way NotifcationViewModelFactory creates view models

Added MessageLimit constant in SmsMessage

 SendSmsViewModel handles SendSmsMessages itself.

It is no longer injected in the Shell, but in the NotificationListViewModel.