[wo-426] Send message to multiple recipients
Added GroupMessageWorkspace that operates on the ContactListViewModel

Styled the SideMenuEntry

Added GroupMessageWorkspace that contains a list of contacts and a Conversation screen

Added ContactListViewModelBase that is implemented by ContactListViewModel and ContactSelectionViewModel

Activates ContactSelection along with its parent

Moved ContactInfoViewModel

Added ContactInfoSelectionViewModel and refactored ContactInfoViewModelFactory

Adds possibility to select contacts when creating a new message

Added GroupMessageViewModel with the possibility to send the same message to all the selected contacts

Adds Recipients label

Added tokenizing control

Added possibility to add contacts from the ContactSelectionViewModel. The newly added contact, is automatically selected and added to the Recipients. Then, the Filter so you can add another contact to the list

Removed TokenizingControl

code cleanup

Adds test for the selection of a newly created contact

Fix merge issues and tests

People and NewMessage screen are both handled by the ContactListViewModel. The difference is that on NewMessage you have multiple selection in the list, and on people, you can only select one contact at a time

Removed unnecessary code. All contacts are handled with the ContactListViewModel

Moved tests from ContactSelectionViewModelTests to ContactListViewModelTests

Added states for ContactInfo

Removed ContactListViewModelBase since it wasn't needed anymore. Moved code inside ContactListViewModel

When sending message to a group of people, a message is saved on each of the recipients.

Added recipients' pictures in the header view

cleanup

Syncing the Listview selected items with the Selected ObservableCollection list

[wo-426] removes cruft

[wo-426] adds resource and prettyfies thinks

[wo-426] removes unused code

[wo-426] makes abstract class abstract and fixes warnings

[wo-426] makes header preatier

[wo-426] removes cruft

[wo-426] updates style name

[wo-426] adds group message translation

[wo-426] moves workspace to proper place